// src/enums.ts
export enum EHeadRotation {
    
    ROT90N = "Rotação -90°",
    ROT45N = "Rotação -45°",
    STOP = "Em Repouso",
    ROT45 = "Rotação 45°",
    ROT90 = "Rotação 90°",
}

export enum EHeadInclination {
    
    UP = "Para Cima",
    STOP = "Em Repouso",
    DOWN = "Para Baixo"
}

export enum EArmElbow {
    
    STOP = "Em Repouso",
    SLIGHTLY_CONTRACTED = "Levemente Contraído",
    CONTRACTED = "Contraído",
    STRONGLY_CONTRACTED = "Fortemente Contraído"
}

export enum EArmPulse {
    ROT90N = "Rotação para -90°",
    ROT45N = "Rotação para -45°",
    STOP = "Em Repouso",
    ROT45 = "Rotação para 45°",
    ROT90 = "Rotação para 90°",
    ROT135 = "Rotação para 135°",
    ROT180 = "Rotação para 180°"
}